pyproject.toml
[tool.poetry]
name = "cg_tokens"
version = "0.1.0"
description = ""
authors = ["Alexander <aleksandr.tugay.02@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
